rewrite ChainStore class
