+++ .gitignore
@@ -4,3 +4,5 @@ npm-debug.log
 node_modules
 dist
 es
+build
+package-lock.json
