+++ lib/chain/index.js
@@ -1,5 +1,5 @@
 import ChainStore from "./src/ChainStore";
-import TransactionBuilder  from "./src/TransactionBuilder";
+import TransactionBuilder from "./src/TransactionBuilder";
 import ChainTypes from "./src/ChainTypes";
 import ObjectId from "./src/ObjectId";
 import NumberUtils from "./src/NumberUtils";
@@ -10,5 +10,15 @@ import Login from "./src/AccountLogin";
 
 const {FetchChainObjects, FetchChain} = ChainStore;
 
-export {ChainStore, TransactionBuilder, FetchChainObjects, ChainTypes,
-    ObjectId, NumberUtils, TransactionHelper, ChainValidation, FetchChain, Login }
+export {
+    ChainStore,
+    TransactionBuilder,
+    FetchChainObjects,
+    ChainTypes,
+    ObjectId,
+    NumberUtils,
+    TransactionHelper,
+    ChainValidation,
+    FetchChain,
+    Login
+};
